health: #184 move blueprint instance & property to schema folder also import from index @ playground tests
