+++ packages/core/src/index.ts
@@ -1,7 +1,5 @@
-export * from "./lib/entity/blueprint/instance";
-export * from "./lib/entity/blueprint/property";
+export * from "./lib/schema/blueprint/instance";
 export * from "./lib/entity/data-structures/complex-key-map";
-export * from "./lib/execution/entity-hydration-query";
 export * from "./lib/entity/data-structures/entity-set";
 export * from "./lib/entity/data-structures/normalized-entities";
 export * from "./lib/entity/entity";
@@ -19,21 +17,23 @@ export * from "./lib/entity/store/entity-store";
 export * from "./lib/entity/store/entity-store-common-index";
 export * from "./lib/entity/store/entity-store-unique-index";
 export * from "./lib/entity/workspace";
-export * from "./lib/expansion/expansion";
-export * from "./lib/query/merge-queries.fn";
-export * from "./lib/query/merge-query.fn";
-export * from "./lib/query/query";
-export * from "./lib/query/reduce-queries.fn";
-export * from "./lib/query/reduce-query.fn";
 export * from "./lib/execution/entity-api";
 export * from "./lib/execution/entity-api-endpoint";
 export * from "./lib/execution/entity-api-endpoint-builder";
+export * from "./lib/execution/entity-hydration-query";
 export * from "./lib/execution/entity-source-gateway";
 export * from "./lib/execution/i-entity-hydrator";
 export * from "./lib/execution/i-entity-source";
 export * from "./lib/execution/query-execution";
 export * from "./lib/execution/query-stream";
 export * from "./lib/execution/query-stream-packet";
+export * from "./lib/expansion/expansion";
+export * from "./lib/query/merge-queries.fn";
+export * from "./lib/query/merge-query.fn";
+export * from "./lib/query/query";
+export * from "./lib/query/reduce-queries.fn";
+export * from "./lib/query/reduce-query.fn";
+export * from "./lib/schema/blueprint/property";
 export * from "./lib/schema/build-default-index-name.fn";
 export * from "./lib/schema/entity-schema";
 export * from "./lib/schema/entity-schema-index";
