+++ packages/core/src/lib/entity/workspace.ts
@@ -23,9 +23,9 @@ import { IEntitySource } from "../execution/i-entity-source";
 import { mergeQueries } from "../query/merge-queries.fn";
 import { Query } from "../query/query";
 import { reduceQueries } from "../query/reduce-queries.fn";
+import { Instance } from "../schema/blueprint/instance";
 import { SchemaCatalog } from "../schema/schema-catalog";
 import { IEntitySchema } from "../schema/schema.interface";
-import { Instance } from "./blueprint/instance";
 import { EntitySet } from "./data-structures/entity-set";
 import { Entity } from "./entity";
 import { createCriterionFromEntities } from "./functions/create-criterion-from-entities.fn";
